Refactor: barrel imports for models and data
