+++ src/server/__tests__/user.test.ts
@@ -1,8 +1,7 @@
 import request from "supertest";
 import app from "../app";
 import knex from "../configs/knex.config.js";
-import deckModel from "../models/deck.model.js";
-import cardModel from "../models/card.model.js";
+import { deckModel, cardModel } from "../models/index.js";
 
 describe("POST /users/register", () => {
   it("returns new user and 201", async () => {
