+++ src/server/data/default-deck.data.ts
@@ -1,6 +1,6 @@
 import { NewDeck } from "../types/deck.js";
 
-export default {
+export const defaultDeck = {
   name: "The Deck of Fire",
   is_scored: false,
   is_custom: false,
