+++ src/server/data/index.ts
@@ -1,0 +1,2 @@
+export { defaultDeck } from "./default-deck.data.js";
+export { defaultCards } from "./default-cards.data.js";
