+++ src/server/models/index.ts
@@ -1,0 +1,3 @@
+export { userModel } from "./user.model.js";
+export { deckModel } from "./deck.model.js";
+export { cardModel } from "./card.model.js";
